Со вторым лвлм чет поделал

+++ Game2/Assets/Scripts/ChaseDeer.cs
@@ -3,8 +3,9 @@ using UnityEngine.SceneManagement;
 
 public class ChaseDeer : MonoBehaviour {
     public Transform EndPoint;
-	public int DeerSpeed;
+	public float DeerSpeed;
     private bool faceRight = true;
+
 
     void Update ()
     {
